ai-interview-coach-backend/backend/requirements.txt
fastapi
uvicorn
numpy
scikit-learn==1.3.2
pandas
nltk
python-multipart
sentence-transformers
torch
spacy>=3.7.0
PyPDF2
python-docx
cohere